fixing scale down issue with Groups and refactoring
